feat(agent-ui): decompose DashboardPage into subcomponents (#242)
- Extracted LeadPipelineSection, FollowUpsSection, PerformanceSection,
  RecentActivitiesSection, QuickActionsSection as independent components
- Moved PIPELINE_COLORS, PIPELINE_LABELS, PRIORITY_BADGE to
  components/dashboard/tokens.ts
- Created useDashboardData hook with useAgentOverview, useAgentPipeline,
  useAgentFollowUps, useAgentPerformance, usePipelineChartData
- Memoized chart data transforms with useMemo in PerformanceSection
  and usePipelineChartData
- DashboardPage.tsx reduced from 821 LOC to 109 LOC
