+++ agent-ui/src/components/dashboard/index.ts
@@ -1,8 +1,13 @@
-export { LeadPipeline } from './LeadPipeline'
-export { TodayFollowUps } from './TodayFollowUps'
-export { RecentActivities } from './RecentActivities'
-export { UpcomingTasks } from './UpcomingTasks'
-export { PerformanceComparison } from './PerformanceComparison'
-export { QuickActions } from './QuickActions'
-export { NotificationsPanel } from './NotificationsPanel'
-export { MyProperties } from './MyProperties'
+export { LeadPipeline } from './LeadPipeline';
+export { LeadPipelineSection } from './LeadPipelineSection';
+export { TodayFollowUps } from './TodayFollowUps';
+export { RecentActivities } from './RecentActivities';
+export { RecentActivitiesSection } from './RecentActivitiesSection';
+export { UpcomingTasks } from './UpcomingTasks';
+export { PerformanceComparison } from './PerformanceComparison';
+export { PerformanceSection } from './PerformanceSection';
+export { QuickActions } from './QuickActions';
+export { QuickActionsSection } from './QuickActionsSection';
+export { NotificationsPanel } from './NotificationsPanel';
+export { MyProperties } from './MyProperties';
+export { FollowUpsSection } from './FollowUpsSection';
